+++ include/kernel/defs.h
@@ -78,6 +78,9 @@ int fork();
 int exit(int exit_code);
 uint64 wait4(int pid, uint64 wstatus, int options);
 uint64 brk(uint64 addr);
+
+// mmap.c
+struct vm_area_struct *find_overlapping_vma(uint64 addr, uint64 len);
 uint64 do_mmap(uint64 addr, uint64 len, int prot, int flags, int fd,
 	       uint64 offset);
 int do_munmap(uint64 addr, uint64 len);